# web/openapi.yaml
openapi: 3.0.3
info:
  title: ICD Converter API
  version: "1.4.0"
  description: >
    API per la conversione e ricerca di codici ICD-9-CM ↔ ICD-10-IM e codici CIPI (nomenclatura italiana). Supporta lookup puntuale, conversione bidirezionale, espansione gerarchica, ricerca testuale full-text e ricerca per similarità semantica con pipeline avanzata: text enrichment dell'indice, query expansion LLM, multi-query splitting e fusione ibrida RRF (embedding + BM25). Include il grouper MS-DRG CMS FY2026 v43.0 con descrizioni in italiano.
servers:
- url: /api/v1
  description: ICD Converter
...
paths:
...
  /icd9/{code}:
    get:
      tags: [ ICD-9-CM ]
      summary: Recupera un codice ICD-9
      description: Restituisce la voce ICD-9-CM corrispondente al codice specificato.
      operationId: getICD9
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "410.1"
        description: Codice ICD-9-CM (es. `410.1`, `250.00`)
      responses:
        "200":
          description: Voce trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ICDEntry"
              example:
                code: "410.1"
                description: "Infarto miocardico acuto della parete anteriore"
                category: "Malattie del sistema circolatorio"
                mappings: [ "I21.0", "I21.09" ]
        "404":
          $ref: "#/components/responses/NotFound"
  /icd9/{code}/to-icd10:
    get:
      tags: [ ICD-9-CM ]
      summary: Converti ICD-9 → ICD-10
      description: Converte un codice ICD-9-CM nei relativi codici ICD-10-IM.
      operationId: icd9ToIcd10
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "410.1"
        description: Codice ICD-9-CM da convertire
      responses:
        "200":
          description: Conversione effettuata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionResponse"
              example:
                source:
                  code: "410.1"
                  description: "Infarto miocardico acuto della parete anteriore"
                  category: "Malattie del sistema circolatorio"
                  mappings: [ "I21.0", "I21.09" ]
                mappings:
                - code: "I21.0"
                  description: "Infarto miocardico acuto transmuratico della parete anteriore"
                  category: "Malattie del sistema circolatorio"
                  mappings: [ "410.1" ]
                - code: "I21.09"
                  description: "Infarto miocardico acuto transmuratico di altra sede della parete anteriore"
                  category: "Malattie del sistema circolatorio"
                  mappings: [ "410.1" ]
        "404":
          $ref: "#/components/responses/NotFound"
  /icd9/{code}/expand:
    get:
      tags: [ ICD-9-CM ]
      summary: Espandi codice padre ICD-9
      description: >
        Restituisce il codice ICD-9 padre insieme a tutti i suoi codici figli (tutti i codici della forma `{code}.xxx`). Utile per esplorare la gerarchia.
      operationId: expandICD9
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "410"
        description: Codice padre ICD-9-CM (es. `410` per ottenere `410.0`, `410.1`, …)
      responses:
        "200":
          description: Espansione effettuata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpandResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /icd9:
    get:
      tags: [ ICD-9-CM ]
      summary: Lista codici ICD-9 (paginata)
      description: Restituisce una lista paginata di tutti i codici ICD-9-CM, con filtro opzionale per categoria.
      operationId: listICD9
      parameters:
      - $ref: "#/components/parameters/page"
      - $ref: "#/components/parameters/limit"
      - $ref: "#/components/parameters/category"
      responses:
        "200":
          description: Lista paginata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"

  # ── ICD-10-IM ─────────────────────────────────────────────────────────────
  /icd10/{code}:
    get:
      tags: [ ICD-10-IM ]
      summary: Recupera un codice ICD-10
      description: Restituisce la voce ICD-10-IM corrispondente al codice specificato.
      operationId: getICD10
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "I21.0"
        description: Codice ICD-10-IM (es. `I21.0`, `E11.9`)
      responses:
        "200":
          description: Voce trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ICDEntry"
              example:
                code: "I21.0"
                description: "Infarto miocardico acuto transmuratico della parete anteriore"
                category: "Malattie del sistema circolatorio"
                mappings: [ "410.1" ]
        "404":
          $ref: "#/components/responses/NotFound"
  /icd10/{code}/to-icd9:
    get:
      tags: [ ICD-10-IM ]
      summary: Converti ICD-10 → ICD-9
      description: Converte un codice ICD-10-IM nei relativi codici ICD-9-CM.
      operationId: icd10ToIcd9
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "I21.0"
        description: Codice ICD-10-IM da convertire
      responses:
        "200":
          description: Conversione effettuata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /icd10/{code}/expand:
    get:
      tags: [ ICD-10-IM ]
      summary: Espandi codice padre ICD-10
      description: >
        Restituisce il codice ICD-10-IM padre insieme a tutti i suoi codici figli. Utile per esplorare la gerarchia (es. `I21` → `I21.0`, `I21.1`, `I21.2`, …).
      operationId: expandICD10
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "I21"
        description: Codice padre ICD-10-IM
      responses:
        "200":
          description: Espansione effettuata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpandResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /icd10:
    get:
      tags: [ ICD-10-IM ]
      summary: Lista codici ICD-10 (paginata)
      description: Restituisce una lista paginata di tutti i codici ICD-10-IM, con filtro opzionale per categoria.
      operationId: listICD10
      parameters:
      - $ref: "#/components/parameters/page"
      - $ref: "#/components/parameters/limit"
      - $ref: "#/components/parameters/category"
      responses:
        "200":
          description: Lista paginata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"

  # ── CIPI ──────────────────────────────────────────────────────────────────
  /cipi/{code}:
    get:
      tags: [ CIPI ]
      summary: Recupera un codice CIPI
      description: Restituisce la voce CIPI corrispondente al codice specificato.
      operationId: getCIPI
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "H1.1"
        description: Codice CIPI (es. `H1.1`, `Q3.1`)
      responses:
        "200":
          description: Voce trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CIPIEntry"
              example:
                code: "H1.1"
                description: "Bypass aortocoronarico (CABG)"
                type: "procedura"
                parent: "H1"
        "404":
          $ref: "#/components/responses/NotFound"
  /cipi/{code}/expand:
    get:
      tags: [ CIPI ]
      summary: Espandi codice padre CIPI
      description: >
        Restituisce il codice CIPI padre insieme a tutti i suoi codici figli diretti (entries con parent = code). Utile per esplorare la gerarchia (es. `H` → `H1`, `H2`, …).
      operationId: expandCIPI
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "H"
        description: Codice padre CIPI
      responses:
        "200":
          description: Espansione effettuata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CIPIExpandResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /cipi:
    get:
      tags: [ CIPI ]
      summary: Lista codici CIPI (paginata)
      description: Restituisce una lista paginata di tutti i codici CIPI, con filtro opzionale per tipo.
      operationId: listCIPI
      parameters:
      - $ref: "#/components/parameters/page"
      - $ref: "#/components/parameters/limit"
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum: [ diagnosi, procedura ]
        description: 'Filtra per tipo: `diagnosi` o `procedura`'
      responses:
        "200":
          description: Lista paginata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CIPIPaginatedResponse"

  # ── Ricerca ───────────────────────────────────────────────────────────────
  /search:
    get:
      tags: [ Ricerca ]
      summary: Ricerca full-text
      description: >
        Ricerca per parola chiave nelle descrizioni ICD-9-CM, ICD-10-IM e/o CIPI. I risultati sono ordinati per punteggio di rilevanza.
      operationId: search
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          example: "infarto miocardico"
        description: Testo da cercare nelle descrizioni
      - name: version
        in: query
        required: false
        schema:
          type: string
          enum: [ icd9, icd10, cipi, both, all ]
          default: both
        description: "Classificazione su cui cercare: `icd9`, `icd10`, `cipi`, `both` (icd9+icd10, default) o `all` (icd9+icd10+cipi)"
      - name: cipi_type
        in: query
        required: false
        schema:
          type: string
          enum: [ diagnosi, procedura ]
        description: "Filtra i risultati CIPI per tipo (applicato solo quando version include `cipi` o `all`)"
      responses:
        "200":
          description: Risultati della ricerca
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
              example:
                query: "infarto miocardico"
                icd9_results:
                - code: "410.1"
                  description: "Infarto miocardico acuto della parete anteriore"
                  category: "Malattie del sistema circolatorio"
                  mappings: [ "I21.0" ]
                  score: 1.0
                icd10_results:
                - code: "I21.0"
                  description: "Infarto miocardico acuto transmuratico della parete anteriore"
                  category: "Malattie del sistema circolatorio"
                  mappings: [ "410.1" ]
                  score: 1.0
                cipi_results: []
        "400":
          $ref: "#/components/responses/BadRequest"
  /search/semantic:
    post:
      tags: [ Ricerca ]
      summary: Ricerca per similarità semantica
      description: >
        Ordina i codici ICD e CIPI per similarità semantica rispetto al testo clinico fornito. Quando `LLM_EMBED_MODEL` è configurato la ricerca usa una pipeline avanzata:

        1. **Query expansion LLM** — se è configurato anche un modello chat (`LLM_API_KEY` + `LLM_MODEL`),
           prima dell'embedding il testo viene inviato all'LLM che produce 2-3 riformulazioni usando
           la terminologia ICD ufficiale italiana (bridging del divario lessicale tra linguaggio clinico
           e terminologia formale).

        2. **Multi-query splitting** — testi clinici lunghi (> 80 caratteri) vengono automaticamente
           spezzati in clausole semanticamente coerenti (su `.` `;` `\n` e `,` per testi molto lunghi).
           Ogni variante (originale + espansioni + clausole) viene embeddata separatamente.

        3. **Fusione ibrida RRF** — i risultati dell'embedding vengono fusi con i risultati BM25
           tramite Reciprocal Rank Fusion (k=60). Garantisce che i match esatti di termini tecnici
           non vengano penalizzati dall'embedding, e che i concetti semantici emergano anche senza
           corrispondenza testuale esatta.

        Se `LLM_EMBED_MODEL` non è configurato, cade in modalità euristica (BM25 keyword scoring).


        Il campo `version` controlla le classificazioni incluse:

        - `all` *(default)*: ICD-9-CM + ICD-10-IM + CIPI

        - `both`: solo ICD-9-CM + ICD-10-IM (retrocompatibile)

        - `icd9`, `icd10`, `cipi`: solo la classificazione specificata


        Il campo `cipi_type` filtra i risultati CIPI:

        - `""` *(default)*: sia diagnosi che procedure CIPI

        - `diagnosi`: solo procedure diagnostiche CIPI

        - `procedura`: solo interventi/procedure terapeutiche CIPI


        Il campo `mode` permette di forzare la modalità:

        - `auto` *(default)*: usa embedding se disponibile, altrimenti euristica

        - `embedding`: richiede l'embedding — restituisce **503** se non configurato

        - `heuristic`: forza sempre la ricerca keyword
      operationId: semanticSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemanticSearchRequest"
            examples:
              infarto:
                summary: Dolore toracico / infarto (tutte le classificazioni)
                value:
                  q: "Paziente con dolore toracico oppressivo, irradiazione al braccio sinistro, sudorazione profusa e nausea. ECG con sopraslivellamento ST in V1-V4. Diagnosi di STEMI anteriore."
                  version: all
                  limit: 5
                  mode: auto
              bypass:
                summary: Bypass coronarico (solo CIPI procedure)
                value:
                  q: "Rivascolarizzazione coronarica mediante bypass aortocoronarico"
                  version: cipi
                  cipi_type: procedura
                  limit: 5
                  mode: auto
              diabete:
                summary: Diabete con complicanze (ICD-10)
                value:
                  q: "Diabete mellito tipo 2 con polineuropatia periferica, ulcera al piede e insufficienza renale cronica stadio 3."
                  version: icd10
                  limit: 10
                  mode: auto
              ictus:
                summary: Ictus ischemico
                value:
                  q: "Paziente con emiplegia destra a insorgenza acuta, afasia, TC cerebrale con area ipodensa in territorio ACM sinistra. Ictus ischemico."
                  version: both
                  limit: 8
                  mode: heuristic
      responses:
        "200":
          description: Risultati ordinati per similarità
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemanticSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          description: Embedding non disponibile (solo con `mode=embedding`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "embedding index not available; set LLM_EMBED_MODEL or use mode=heuristic"

  # ── DRG ─────────────────────────────────────────────────────────────────
  /drg/{code}:
    get:
      tags: [ DRG ]
      summary: Lookup singolo codice MS-DRG
      description: Restituisce i dettagli di un codice MS-DRG (descrizione italiana, tipo, peso relativo, degenze attese, MDC di appartenenza).
      operationId: getDRG
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "280"
        description: Codice DRG numerico, con o senza zeri iniziali (es. `1`, `001`, `280`, `470`).
      responses:
        "200":
          description: DRG trovato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DRGEntry"
              example:
                code: "280"
                description: "INFARTO MIOCARDICO ACUTO, DIMESSO VIVO CON CMM"
                type: "MED"
                mdc: "05"
                mdc_description: "Malattie e Disturbi del Sistema Circolatorio"
                weight: 2.1234
                geometric_los: 4.1
                arithmetic_los: 5.2
        "404":
          $ref: "#/components/responses/NotFound"
  /drg/search:
    get:
      tags: [ DRG ]
      summary: Ricerca testuale DRG
      description: Ricerca full-text nelle descrizioni MS-DRG con filtri opzionali per MDC e tipo (SURG/MED).
      operationId: searchDRG
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          example: "insufficienza cardiaca"
        description: Testo da ricercare nelle descrizioni DRG.
      - name: mdc
        in: query
        required: false
        schema:
          type: string
          example: "05"
        description: Filtra per codice MDC (es. `05`).
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum: [ SURG, MED ]
          example: "MED"
        description: Filtra per tipo — `SURG` (chirurgico) o `MED` (medico).
      - $ref: "#/components/parameters/page"
      - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Lista DRG corrispondenti
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 12
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/DRGEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /mdc/{code}:
    get:
      tags: [ DRG ]
      summary: Dettaglio MDC con DRG
      description: Restituisce i dettagli di una MDC e tutti i DRG ad essa associati.
      operationId: getMDC
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: "05"
        description: Codice MDC a due cifre (es. `05`, `12`, `PRE`).
      responses:
        "200":
          description: MDC trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MDCDetail"
              example:
                code: "05"
                description: "Malattie e Disturbi del Sistema Circolatorio"
                drg_count: 67
                drgs:
                - code: "280"
                  description: "INFARTO MIOCARDICO ACUTO, DIMESSO VIVO CON CMM"
                  type: "MED"
                  weight: 2.1234
                  geometric_los: 4.1
                  arithmetic_los: 5.2
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  schemas:
    ICDEntry:
      type: object
      description: Voce ICD con codice, descrizione, categoria e mappings cross-version
      properties:
        code:
          type: string
          description: Codice ICD (es. `410.1` per ICD-9, `I21.0` per ICD-10)
          example: "410.1"
        description:
          type: string
          description: Descrizione clinica del codice
          example: "Infarto miocardico acuto della parete anteriore"
        category:
          type: string
          description: Capitolo/categoria di appartenenza
          example: "Malattie del sistema circolatorio"
        mappings:
          type: array
          description: Codici equivalenti nell'altra versione ICD
          items:
            type: string
          example: [ "I21.0", "I21.09" ]
...
    ConversionResponse:
      type: object
      description: Risultato di una conversione ICD-9 ↔ ICD-10
      properties:
        source:
          $ref: "#/components/schemas/ICDEntry"
        mappings:
          type: array
          description: Codici equivalenti nella versione di destinazione
          items:
            $ref: "#/components/schemas/ICDEntry"
    ExpandResponse:
      type: object
      description: Codice padre con tutti i suoi codici figli (espansione gerarchica)
      properties:
        parent:
          $ref: "#/components/schemas/ICDEntry"
        children:
          type: array
          description: Sottocodici del codice padre
          items:
            $ref: "#/components/schemas/ICDEntry"
    CIPIExpandResponse:
      type: object
      description: Codice CIPI padre con tutti i suoi figli diretti
      properties:
        parent:
          $ref: "#/components/schemas/CIPIEntry"
        children:
          type: array
          description: Codici figli diretti (entries con parent = codice padre)
          items:
            $ref: "#/components/schemas/CIPIEntry"
...
  responses:
    NotFound:
      description: Codice ICD non trovato
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "ICD-9 code not found: 999.9"
...
  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Numero di pagina (1-based)
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
      description: Risultati per pagina (max 200)
    category:
      name: category
      in: query
      required: false
      schema:
        type: string
      description: 'Filtra per categoria/capitolo (es. "Malattie del sistema circolatorio")'